refactored code into something more usable for the API
